Split up app and server to enable testing
